datafix/correct/src/divergences/feebee/misc/requirements.txt
tensorflow_hub==0.12.0
tensorflow-gpu==2.4.0
torch==1.7.1+cu110
absl-py==0.10.0
sklearn==0.22.2
seaborn==0.11.2
